requirements.txt
ezregex>=1.4.3
streamlit
streamlit-code-editor
